debug the actual http requests' method and URI
There already is a debug statement for the response, however without also
logging the requests that information is only partially useful. In
particular when a higher level piece of software using the API runs into
problems it will often be easier to debug backwards from a request, which
kind of requires knowing what the request was.